+++ hooks/index.ts
@@ -2,3 +2,6 @@ export { useGalleryNavigation } from './useGalleryNavigation';
 export { useAudioController } from './useAudioController';
 export { useDeviceCapabilities } from './useDeviceCapabilities';
 export type { UseDeviceCapabilitiesReturn } from './useDeviceCapabilities';
+export { useSpatialAudio } from './useSpatialAudio';
+export { useAnalytics } from './useAnalytics';
+export type { AnalyticsEventType } from './useAnalytics';